Changing objects mutations location
